chore: make match-mirror standalone
